+++ backend/canisters/community/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Added
+
+- Support mentioning user groups ([#4308](https://github.com/open-chat-labs/open-chat/pull/4308))
+
 ## [[2.0.827](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.827-community)] - 2023-09-01
 
 ### Added